requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
streamlit>=1.28.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
plotly>=5.18.0
pyyaml>=6.0
joblib>=1.3.0
pytest>=8.2.0
